Custom DocumentListener for for HexSearch and exception handling
