+++ runtime-decompiler/src/main/java/org/fife/ui/hex/event/HexSearchDocumentListener.java
@@ -1,0 +1,47 @@
+package org.fife.ui.hex.event;
+
+import org.fife.ui.hex.swing.HexEditor;
+import org.fife.ui.hex.swing.HexSearch;
+import org.jrd.backend.core.OutputController;
+
+import javax.swing.*;
+import javax.swing.event.DocumentEvent;
+import javax.swing.event.DocumentListener;
+
+public class HexSearchDocumentListener implements DocumentListener {
+
+    private final HexEditor hex;
+    private final HexSearch hexSearchEngine;
+    private final JTextField hexSearch;
+    private final JComboBox<HexSearch.HexSearchOptions> hexSearchType;
+
+    public HexSearchDocumentListener(HexEditor hex, HexSearch hexSearchEngine, JTextField hexSearch, JComboBox<HexSearch.HexSearchOptions> hexSearchType) {
+        this.hex = hex;
+        this.hexSearchEngine = hexSearchEngine;
+        this.hexSearch = hexSearch;
+        this.hexSearchType = hexSearchType;
+    }
+
+    @Override
+    public void insertUpdate(DocumentEvent documentEvent) {
+        find();
+    }
+
+    @Override
+    public void removeUpdate(DocumentEvent documentEvent) {
+        find();
+    }
+
+    @Override
+    public void changedUpdate(DocumentEvent documentEvent) {
+        find();
+    }
+
+    private void find() {
+        try {
+            hexSearchEngine.searchHexCode(hexSearch.getText(), (HexSearch.HexSearchOptions) hexSearchType.getSelectedItem());
+        } catch (Exception e) {
+            OutputController.getLogger().log(e);
+        }
+    }
+}
